Fix  read_stb function which broke after earlier commit
The usbtmc488_ioctl_read_stb relied on a positive return from
usbtmc_get_stb to reset the srq condition in the driver.
Commit c79080c1d9bb ("Fix ioctl return value aligned with kernel version")
changed the return value of usbtmc_get_stb to 0 on success instead of
returning the control value of usb_control_msg which is positive in
the normal case. This change caused usbtmc488_ioctl_read_stb to no
longer function correctly.

Change the test in usbtmc488_ioctl_read_stb to test for failure
first and return the failure code immediately.